Fixing all the JS features beug
