+++ app/assets/stylesheets/_searchbar.scss
@@ -1,6 +1,9 @@
 #results {
-    font-size: $body_size;
-    font-family: $sl-font-family--body;
+  font-size: $body_size;
+  font-family: $sl-font-family--body;
+  position: absolute;
+  background-color: white;
+  width: 100%;
 
   & .avatar {
     margin-left: -30px;
