+++ app/javascript/packs/application.js
@@ -41,6 +41,8 @@ dislike.addEventListener('click', (event) => {
 import { showTabContentOnClick } from '../components/tabs_profile';
 import { shuffleMovie } from '../components/movie_shuffle';
 import { fetchMovies } from './movies';
+import { filterCategories } from '../components/filter_categories';
+import { searchBar } from '../components/search_bar';
 // import { followButton } from '../components/follow_button_profile';
 
 const input = document.querySelector('#search');
@@ -52,5 +54,5 @@ const movieIndex = document.querySelector('#search');
 // shuffleMovie(shuffle);
 
 showTabContentOnClick();
-
-
+filterCategories();
+searchBar();
