+++ app/views/dashboards/index.html.erb
@@ -11,7 +11,7 @@
 <!-- DO NOT REMOVE THE CLASS CARDS OTHERWISE IT FUCK EVERYTHING -->
 <div class="cards">
   <% @tastes.reverse.each do |taste| %>
-    <div class="dashboard-review-container flex-column">
+    <div class="dashboard-review-container flex-column" data-genre="<%= taste.movie.genre %>">
       <div class="section-card d-flex justify-content-between">
         <div class="review-text d-flex">
           <%= link_to user_path(taste.user) do %>
